backend/requirements.txt
fastapi[standard]
pydantic
uvicorn
requests
psycopg[binary]
psycopg[pool]
pwdlib[argon2]
pyjwt